moved qpi hash map to qpi.h and qpi_hash_map_impl.h in contract_core
