+++ src/Qubic.vcxproj.filters
@@ -182,6 +182,9 @@
     <ClInclude Include="platform\assert.h">
       <Filter>platform</Filter>
     </ClInclude>
+    <ClInclude Include="..\test\qpi_hash_map_impl.h">
+      <Filter>contract_core</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="platform">
