+++ test/test.vcxproj.filters
@@ -20,8 +20,6 @@
   <ItemGroup>
     <ClInclude Include="score_reference.h" />
     <ClInclude Include="score_params.h" />
-    <ClInclude Include="qpi_hash_map.h" />
-    <ClInclude Include="qpi_hash_map_impl.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
